move and remove contents of runner submodule into api.rs
