+++ crates/tcli/src/package/install/api.rs
@@ -2,7 +2,14 @@ use std::path::PathBuf;
 
 use serde::{Deserialize, Serialize};
 
-use crate::package::Package;
+use crate::{package::Package, ts::version::Version};
+
+/// This is the minimum support 
+pub static INSTALLER_VERSION: Version = Version {
+    major: 1,
+    minor: 0,
+    patch: 0,
+};
 
 /// Arguments are passed into the installer executable as a JSON string, not by argument
 /// name-value pairs. This means that the installer's dev can rely on JSON deserialization
@@ -21,24 +28,25 @@ pub enum Request {
         game_dir: PathBuf,
         tracked_files: Vec<PathBuf>,
     },
+    StartGame {
+        mods_enabled: bool,
+        project_state: PathBuf,
+        game_dir: PathBuf,
+        game_exe: PathBuf,
+    },
 }
 
 #[derive(Serialize, Deserialize)]
-pub struct ArgHeader {
-    version: u32,
-    payload: ArgVariants,
-}
-
-#[derive(Serialize, Deserialize)]
-pub enum ArgVariants {
-    InstallerVersion,
-    InstallArgs(InstallerArgs),
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct InstallerArgs {
-    arg_version: u32,
-    package: Package,
-    package_dir: PathBuf,
-    game_dir: PathBuf,
+#[serde(tag = "type", content = "payload")]
+pub enum Response {
+    Version {
+        installer_version: Version,
+        protocol_version: Version,
+    },
+    PackageInstall {
+        tracked_files: Vec<PathBuf>,
+    },
+    PackageUninstall {
+        tracked_files: Vec<PathBuf>,
+    },
 }
